Split autocomplete into two files with a shared helper file.  No actual real changes at this point.
